fix: extract form in component
