+++ src/app/(private)/hunt/list/page.tsx
@@ -1,0 +1,8 @@
+import ListHuntView from "@/ui/pages/hunt/list";
+
+async function ListHuntsPage() {
+  // fazer um fetch baseado no usuário autenticado
+  return <ListHuntView />;
+}
+
+export default ListHuntsPage;
