+++ src/ui/pages/hunt/list.tsx
@@ -1,0 +1,5 @@
+const ListHuntView = () => {
+  return <>Listagens de hunts</>;
+};
+
+export default ListHuntView;
